move psi executor from app_common to app_opt (#1372)
* move psi executor from app_common to app_opt

* 1) update format
2) Unit test adding additional optional dependency install

* 1) update format
2) Unit test adding additional optional dependency install
3) exclude Tenseal
